+++ src/controller/index.js
@@ -12,7 +12,7 @@ export default class extends base {
     }
     //控制器默认方法
     indexAction () {
-        this.set('currentNav', 'index');
+        this.assign('currentNav', 'index');
         return this.render();
     }
 }